remove config source
