+++ crates/settings/src/unit_test.rs
@@ -1,10 +1,13 @@
+use crate::blocks::BlocksCfg;
 use crate::*;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::sync::{Arc, RwLock};
 use strict_yaml_rust::StrictYaml;
 #[cfg(target_family = "wasm")]
-use wasm_bindgen_utils::{impl_wasm_traits, prelude::*};
+use wasm_bindgen_utils::{
+    impl_wasm_traits, prelude::*, serialize_hashmap_as_object, serialize_opt_hashmap_as_object,
+};
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
 #[serde(rename_all = "kebab-case")]
@@ -14,6 +17,34 @@ pub struct UnitTestConfigSource {
 }
 #[cfg(target_family = "wasm")]
 impl_wasm_traits!(UnitTestConfigSource);
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+#[serde(rename_all = "kebab-case")]
+#[cfg_attr(target_family = "wasm", derive(Tsify))]
+pub struct ScenarioConfigSource {
+    #[serde(default, skip_serializing_if = "HashMap::is_empty")]
+    #[cfg_attr(
+        target_family = "wasm",
+        serde(serialize_with = "serialize_hashmap_as_object"),
+        tsify(optional, type = "Record<string, string>")
+    )]
+    pub bindings: HashMap<String, String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub runs: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub blocks: Option<BlocksCfg>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub deployer: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[cfg_attr(
+        target_family = "wasm",
+        serde(serialize_with = "serialize_opt_hashmap_as_object"),
+        tsify(optional, type = "Record<string, ScenarioConfigSource>")
+    )]
+    pub scenarios: Option<HashMap<String, ScenarioConfigSource>>,
+}
+#[cfg(target_family = "wasm")]
+impl_wasm_traits!(ScenarioConfigSource);
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
 #[serde(rename_all = "kebab-case")]
